lib: merge onread handlers for http2 streams & net.Socket
