+++ lib/internal/http2/core.js
@@ -113,7 +113,9 @@ const {
 const {
   createWriteWrap,
   writeGeneric,
-  writevGeneric
+  writevGeneric,
+  _onread,
+  kUpdateTimer
 } = require('internal/stream_base_commons');
 
 const { ShutdownWrap } = process.binding('stream_wrap');
@@ -146,7 +148,6 @@ const kServer = Symbol('server');
 const kSession = Symbol('session');
 const kState = Symbol('state');
 const kType = Symbol('type');
-const kUpdateTimer = Symbol('update-timer');
 
 const kDefaultSocketTimeout = 2 * 60 * 1000;
 
@@ -361,31 +362,10 @@ function onStreamClose(code) {
 // Receives a chunk of data for a given stream and forwards it on
 // to the Http2Stream Duplex for processing.
 function onStreamRead(nread, buf) {
+  const handle = this;
   const stream = this[kOwner];
-  if (nread >= 0 && !stream.destroyed) {
-    debug(`Http2Stream ${stream[kID]} [Http2Session ` +
-          `${sessionName(stream[kSession][kType])}]: receiving data chunk ` +
-          `of size ${nread}`);
-    stream[kUpdateTimer]();
-    if (!stream.push(buf)) {
-      if (!stream.destroyed)  // we have to check a second time
-        this.readStop();
-    }
-    return;
-  }
 
-  // Last chunk was received. End the readable side.
-  debug(`Http2Stream ${stream[kID]} [Http2Session ` +
-        `${sessionName(stream[kSession][kType])}]: ending readable.`);
-
-  // defer this until we actually emit end
-  if (stream._readableState.endEmitted) {
-    stream[kMaybeDestroy]();
-  } else {
-    stream.on('end', stream[kMaybeDestroy]);
-    stream.push(null);
-    stream.read(0);
-  }
+  _onread(handle, stream, nread, buf);
 }
 
 // Called when the remote peer settings have been updated.
@@ -2122,6 +2102,7 @@ function afterOpen(session, options, headers, streamOptions, err, fd) {
 class ServerHttp2Stream extends Http2Stream {
   constructor(session, handle, id, options, headers) {
     super(session, options);
+    handle.owner = this;
     this[kInit](id, handle);
     this[kProtocol] = headers[HTTP2_HEADER_SCHEME];
     this[kAuthority] = headers[HTTP2_HEADER_AUTHORITY];
