+++ lib/internal/stream_base_commons.js
@@ -1,10 +1,10 @@
 'use strict';
 
 const { Buffer } = require('buffer');
-const errors = require('internal/errors');
 const { WriteWrap } = process.binding('stream_wrap');
-
-const errnoException = errors.errnoException;
+const { UV_EOF } = process.binding('uv');
+const { errnoException } = require('internal/errors');
+const kUpdateTimer = Symbol('update-timer');
 
 function handleWriteReq(req, data, encoding) {
   const { handle } = req;
@@ -81,8 +81,40 @@ function afterWriteDispatched(self, req, err, cb) {
   }
 }
 
+function _onread(handle, stream, nread, buf) {
+  stream[kUpdateTimer]();
+
+  if (nread > 0 && !stream.destroyed) {
+    const ret = stream.push(buf);
+    if (handle.reading && !ret) {
+      handle.reading = false;
+      const err = handle.readStop();
+      if (err && stream.destroy)
+        stream.destroy(errnoException(err, 'read'));
+    }
+
+    return;
+  }
+
+  if (nread === 0) {
+    return;
+  }
+
+  if (nread !== UV_EOF) {
+    return stream.destroy(errnoException(nread, 'read'));
+  }
+
+  // push a null to signal the end of data.
+  // Do it before `maybeDestroy` for correct order of events:
+  // `end` -> `close`
+  stream.push(null);
+  stream.read(0);
+}
+
 module.exports = {
   createWriteWrap,
   writevGeneric,
-  writeGeneric
+  writeGeneric,
+  _onread,
+  kUpdateTimer
 };
