+++ lib/net.js
@@ -36,8 +36,7 @@ const {
 const assert = require('assert');
 const {
   UV_EADDRINUSE,
-  UV_EINVAL,
-  UV_EOF
+  UV_EINVAL
 } = process.binding('uv');
 
 const { Buffer } = require('buffer');
@@ -61,7 +60,9 @@ const {
 const {
   createWriteWrap,
   writevGeneric,
-  writeGeneric
+  writeGeneric,
+  _onread,
+  kUpdateTimer
 } = require('internal/stream_base_commons');
 const errors = require('internal/errors');
 const {
@@ -513,6 +514,12 @@ Object.defineProperty(Socket.prototype, 'bufferSize', {
   }
 });
 
+Object.defineProperty(Socket.prototype, kUpdateTimer, {
+  get: function() {
+    return this._unrefTimer;
+  }
+});
+
 
 // Just call handle.readStart until we have enough in the buffer
 Socket.prototype._read = function(n) {
@@ -619,53 +626,9 @@ Socket.prototype._destroy = function(exception, cb) {
 // buffer, or when there's an error reading.
 function onread(nread, buffer) {
   var handle = this;
-  var self = handle.owner;
-  assert(handle === self._handle, 'handle != self._handle');
+  var stream = handle.owner;
 
-  self._unrefTimer();
-
-  debug('onread', nread);
-
-  if (nread > 0) {
-    debug('got data');
-
-    // read success.
-    // In theory (and in practice) calling readStop right now
-    // will prevent this from being called again until _read() gets
-    // called again.
-
-    // Optimization: emit the original buffer with end points
-    var ret = self.push(buffer);
-
-    if (handle.reading && !ret) {
-      handle.reading = false;
-      debug('readStop');
-      var err = handle.readStop();
-      if (err)
-        self.destroy(errnoException(err, 'read'));
-    }
-    return;
-  }
-
-  // if we didn't get any bytes, that doesn't necessarily mean EOF.
-  // wait for the next one.
-  if (nread === 0) {
-    debug('not any data, keep waiting');
-    return;
-  }
-
-  // Error, possibly EOF.
-  if (nread !== UV_EOF) {
-    return self.destroy(errnoException(nread, 'read'));
-  }
-
-  debug('EOF');
-
-  // push a null to signal the end of data.
-  // Do it before `maybeDestroy` for correct order of events:
-  // `end` -> `close`
-  self.push(null);
-  self.read(0);
+  _onread(handle, stream, nread, buffer);
 }
